refactor(intent): move tests to internal/intent package (GH-812)
Move anthropic_test.go and conversation_test.go from
internal/adapters/telegram/ to internal/intent/ to co-locate
tests with their canonical implementations.

Also removes duplicate Intent type declarations from classifier.go
that were already defined in intent.go.
